First commit of WIP.
This code is embarrassing, but I'm going to check it in anyway. Otherwise, there will only be a single commit at the end, which doesn't help anyone evaluate my incremental development approach.
At this point, I have a CLI tool that can almost import Projects, Issues, and Labels idempotently.
There is tonnes of code duplication, a complete lack of error handling, and a whole lot more functionality and polish to go.